Trait scad_runtime::Callbacks

pub trait Callbacks {
    fn before_execute(
        &mut self,
        _current_chunk: &Chunk,
        _instruction_pointer: usize,
        _instruction: Instruction,
        _stack: &Stack
    ) { ... }
    fn consume_geometry(&mut self, _geometry: Geometry) { ... }
    fn print(&mut self, _values: &[Value]) -> Result<(), RuntimeError> { ... }
}

Provided Methods

fn before_execute(
&mut self,
_current_chunk: &Chunk,
_instruction_pointer: usize,
_instruction: Instruction,
_stack: &Stack
)

A callback that is fired immediately before executing the next instruction.

At this point, the instruction pointer is still pointing at the current instruction.

fn consume_geometry(&mut self, _geometry: Geometry)

fn print(&mut self, _values: &[Value]) -> Result<(), RuntimeError>